update test names, move tests
