+++ ui/components/app/snaps/snap-ui-renderer/snap-ui-renderer.test.js
@@ -6,10 +6,23 @@ import {
   Button,
   Input,
 } from '@metamask/snaps-sdk/jsx';
+import { fireEvent } from '@testing-library/react';
 import { BackgroundColor } from '../../../../helpers/constants/design-system';
-import { renderInterface } from './test-utils';
+import * as backgroundConnection from '../../../../store/background-connection';
+import { renderInterface, MOCK_INTERFACE_ID, MOCK_SNAP_ID } from './test-utils';
+
+jest.mock('../../../../store/background-connection', () => ({
+  ...jest.requireActual('../../../../store/background-connection'),
+  submitRequestToBackground: jest.fn(),
+}));
+
+const { submitRequestToBackground } = jest.mocked(backgroundConnection);
 
 describe('SnapUIRenderer', () => {
+  beforeEach(() => {
+    jest.resetAllMocks();
+  });
+
   it('renders loading state', () => {
     const { container } = renderInterface(null);
 
@@ -68,6 +81,72 @@ describe('SnapUIRenderer', () => {
         'mm-box snap-ui-renderer__content mm-box--background-color-background-default',
       ),
     ).toHaveLength(1);
+    expect(container).toMatchSnapshot();
+  });
+
+  it('renders footers', () => {
+    const { container, getByText } = renderInterface(
+      Container({
+        children: [
+          Box({ children: Text({ children: 'Hello world!' }) }),
+          Footer({ children: Button({ children: 'Foo' }) }),
+        ],
+      }),
+      { useFooter: true },
+    );
+
+    expect(getByText('Foo')).toBeDefined();
+    expect(container).toMatchSnapshot();
+  });
+
+  it('supports interactive inputs', () => {
+    const { container, getByRole } = renderInterface(
+      Box({ children: Input({ name: 'input' }) }),
+    );
+
+    const input = getByRole('textbox');
+    fireEvent.change(input, { target: { value: 'a' } });
+
+    expect(submitRequestToBackground).toHaveBeenNthCalledWith(
+      1,
+      'updateInterfaceState',
+      [MOCK_INTERFACE_ID, { input: 'a' }],
+    );
+
+    expect(submitRequestToBackground).toHaveBeenNthCalledWith(
+      2,
+      'handleSnapRequest',
+      [
+        {
+          handler: 'onUserInput',
+          origin: '',
+          request: {
+            jsonrpc: '2.0',
+            method: ' ',
+            params: {
+              context: null,
+              event: { name: 'input', type: 'InputChangeEvent', value: 'a' },
+              id: MOCK_INTERFACE_ID,
+            },
+          },
+          snapId: MOCK_SNAP_ID,
+        },
+      ],
+    );
+
+    expect(container).toMatchSnapshot();
+  });
+
+  it('prefills interactive inputs with existing state', () => {
+    const { container, getByRole } = renderInterface(
+      Box({ children: Input({ name: 'input' }) }),
+      { state: { input: 'bar' } },
+    );
+
+    const input = getByRole('textbox');
+    expect(input).toBeDefined();
+    expect(input.value).toStrictEqual('bar');
+
     expect(container).toMatchSnapshot();
   });
